+++ loader/src/loader/Dirs.cpp
@@ -3,37 +3,21 @@
 #include <cocos2d.h>
 #include <crashlog.hpp>
 #include <filesystem>
+#include "LoaderImpl.hpp"
 
 using namespace geode::prelude;
 
 ghc::filesystem::path dirs::getGameDir() {
-    return ghc::filesystem::path(CCFileUtils::sharedFileUtils()->getWritablePath2().c_str());
-}
-
-ghc::filesystem::path dirs::getSaveDir() {
-#ifdef GEODE_IS_MACOS
-    // not using ~/Library/Caches
-    return ghc::filesystem::path("/Users/Shared/Geode");
-#elif defined(GEODE_IS_WINDOWS)
-    // this is std::filesystem intentionally because ghc version doesnt want to work with softlinked directories
-    return ghc::filesystem::path(
-        std::filesystem::weakly_canonical(CCFileUtils::sharedFileUtils()->getWritablePath().c_str())
-            .string()
-    );
-#else
-    // return ghc::filesystem::path(CCFileUtils::sharedFileUtils()->getWritablePath().c_str());
-    return ghc::filesystem::path("/data/data/com.geode.launcher");
-#endif
+    #ifdef GEODE_IS_WINDOWS
+        // this is std::filesystem intentionally because ghc version doesnt want to work with softlinked directories
+        return std::filesystem::weakly_canonical(path.string()).string();
+    #else
+        return ghc::filesystem::weakly_canonical(path.string()).string();
+    #endif
 }
 
 ghc::filesystem::path dirs::getGeodeDir() {
-#ifdef GEODE_IS_MACOS
-    return ghc::filesystem::current_path() / "geode";
-#elif defined(GEODE_IS_ANDROID)
-    return dirs::getSaveDir() / "ggeode";
-#else
     return dirs::getGameDir() / "geode";
-#endif
 }
 
 ghc::filesystem::path dirs::getGeodeSaveDir() {