+++ loader/src/platform/windows/util.cpp
@@ -4,11 +4,12 @@
 #ifdef GEODE_IS_WINDOWS
 
 using namespace geode::prelude;
-
+#include <Geode/loader/Dirs.hpp>
 #include "nfdwin.hpp"
 #include <ghc/fs_fwd.hpp>
 #include <Windows.h>
 #include <iostream>
+#include <ShlObj.h>
 #include <shlwapi.h>
 #include <shobjidl.h>
 #include <sstream>
@@ -113,6 +114,46 @@ CCPoint cocos::getMousePos() {
     auto frameSize = gl->getFrameSize();
     auto mouse = gl->getMousePosition() / frameSize;
     return ccp(mouse.x, 1.f - mouse.y) * winSize;
+}
+
+ghc::filesystem::path dirs::getGameDir() {
+    // only fetch the path once, since ofc it'll never change
+    // throughout the execution
+    static const auto path = [] {
+        std::array<WCHAR, MAX_PATH> buffer;
+        GetModuleFileNameW(NULL, buffer.data(), MAX_PATH);
+
+        const ghc::filesystem::path path(buffer.data());
+        return path.parent_path();
+    }();
+
+    return path;
+}
+
+ghc::filesystem::path dirs::getSaveDir() {
+    // only fetch the path once, since ofc it'll never change
+    // throughout the execution
+    static const auto path = [] {
+        std::array<WCHAR, MAX_PATH + 1> buffer;
+        GetModuleFileNameW(NULL, buffer.data(), MAX_PATH + 1);
+
+        auto executablePath = ghc::filesystem::path(buffer.data());
+        auto executableName = executablePath.filename().wstring();
+        executableName = executableName.substr(0, executableName.find_last_of(L"."));
+
+        if (SHGetFolderPathW(NULL, CSIDL_LOCAL_APPDATA, NULL, SHGFP_TYPE_CURRENT, buffer.data()) >= 0) {
+            auto appdataPath = ghc::filesystem::path(buffer.data());
+            auto savePath = appdataPath / executableName;
+
+            if (SHCreateDirectoryExW(NULL, savePath.wstring().c_str(), NULL) >= 0) {
+                return savePath;
+            }
+        }
+        
+        return executablePath.parent_path();
+    }();
+
+    return path;
 }
 
 #endif